修改toInt方法

+++ src/main/java/com/lorne/core/framework/utils/math/MathDecimal.java
@@ -52,7 +52,7 @@ public class MathDecimal {
         return count.doubleValue();
     }
 
-    public double toInt(){
+    public int toInt(){
         return count.intValue();
     }
